machine-hub-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
wakeonlan==3.1.0